refactored code with its own route, controller and get function
